+++ test/modules/additive.rb
@@ -1,0 +1,13 @@
+export :foo
+export :bar
+export :BAZ
+
+def foo
+  :foo
+end
+
+def bar
+  :bar
+end
+
+BAZ = :baz