+++ test/modules/export_with_export_default.rb
@@ -1,0 +1,8 @@
+export :foo
+export_default :BAZ
+
+def foo
+  :bar
+end
+
+BAZ = [:baz]