+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+* Implement additive exports
 * Refactor auto_import_map, add not_found option
 * Fix backtrace for exports of missing symbols